// client/src/types.ts
export enum Eixo {
  HUMANAS = 'Ciências Humanas e Sociais Aplicadas',
  NATUREZA = 'Ciências da Natureza e suas Tecnologias',
  LINGUAGENS_1 = 'Linguagens e suas Tecnologias - 1 (Arte, Língua Portuguesa e Redação)',
  LINGUAGENS_2 = 'Linguagens e suas Tecnologias - 2 (Inglês)',
  LINGUAGENS_3 = 'Linguagens e suas Tecnologias - 3 (Educação Física)',
  MATEMATICA = 'Matemática e suas Tecnologias',
  COMPUTACAO = 'Programação e Pensamento Computacional',
}

export enum Subject {
  ARTE = 'Arte',
  BIOLOGIA = 'Biologia',
  GEOGRAFIA = 'Geografia',
  HISTORIA = 'História',
  FILOSOFIA = 'Filosofia',
  SOCIOLOGIA = 'Sociologia',
  INGLES = 'Língua Inglesa',
  PORTUGUES = 'Língua Portuguesa',
  REDACAO = 'Redação',
  MATEMATICA = 'Matemática',
  FISICA = 'Física',
  QUIMICA = 'Química',
  PROJETO_VIDA = 'Projetos de Vida',
  PENSAMENTO_COMPUTACIONAL = 'Pensamento Computacional',
  PROGRAMACAO = 'Programação',
  EDUCACAO_FISICA = 'Educação Física',
}

export interface Book {
  code: string;
  title: string;
  publisher: string;
}

export interface Vote {
  textbook1: string; // book code
  textbook2: string; // book code
  project1: string; // book code
  project2: string; // book code
}

export interface Teacher {
  id: string;
  name: string;
  subjects: Subject[];
  eixo: Eixo;
  vote?: Vote;
}